tiemcuagio_django/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
Django==5.1.2
django-allauth==65.0.2
django-ckeditor==6.7.1
django-ckeditor-5==0.2.15
django-cors-headers==4.6.0
django-extensions==3.2.3
django-js-asset==2.2.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
gunicorn==23.0.0
idna==3.10
packaging==24.1
pillow==11.0.0
pycparser==2.22
PyJWT==2.9.0
pytz==2024.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
uuid==1.30
